initial project structure refactorization
