+++ data/test_instance.csv
@@ -1,0 +1,127 @@
+demand;;;
+1;4;3;2
+2;3;4;2
+3;3;3;4
+4;4;3;2
+5;3;2;2
+6;3;2;3
+7;2;2;2
+8;4;2;2
+9;2;2;4
+10;3;4;2
+11;4;2;4
+12;2;3;2
+13;2;3;4
+14;2;4;2
+15;2;3;4
+16;4;2;3
+17;3;2;4
+18;3;2;2
+19;3;3;3
+20;2;4;3
+21;3;2;3
+22;3;3;4
+23;3;3;2
+24;4;3;3
+25;3;4;3
+26;3;2;3
+27;2;2;3
+28;2;2;2
+workhours;;;
+1;160;;
+2;80;;
+3;40;;
+4;80;;
+5;160;;
+6;160;;
+7;160;;
+8;160;;
+9;160;;
+10;160;;
+11;160;;
+12;160;;
+13;80;;
+14;160;;
+15;160;;
+vacation;;;
+1;21;;
+2;7;;
+3;28;;
+4;9;;
+5;5;;
+6;9;;
+6;25;;
+7;6;;
+8;12;;
+9;7;;
+10;22;;
+11;12;;
+12;8;;
+13;24;;
+14;7;;
+14;22;;
+15;1;;
+preferred_companions;;;
+1;7;;
+2;7;;
+3;6;;
+4;9;;
+5;5;;
+6;11;;
+7;9;;
+8;15;;
+9;10;;
+10;4;;
+11;2;;
+12;1;;
+13;6;;
+14;13;;
+15;10;;
+unpreferred_companions;;;
+1;6;;
+2;11;;
+3;14;;
+4;5;;
+5;10;;
+6;8;;
+7;13;;
+8;3;;
+9;15;;
+10;5;;
+11;6;;
+12;10;;
+13;8;;
+14;6;;
+15;4;;
+preferred_shifts;;;
+1;5;3;
+2;5;1;
+3;27;1;
+4;22;1;
+5;21;1;
+6;14;3;
+7;26;2;
+8;15;3;
+9;5;1;
+10;21;1;
+11;27;1;
+12;15;2;
+13;11;3;
+14;13;1;
+15;10;3;
+unpreferred_shifts;;;
+1;2;3;
+2;22;3;
+3;21;1;
+4;8;1;
+5;9;2;
+6;2;3;
+7;11;1;
+8;22;2;
+9;16;1;
+10;10;3;
+11;8;3;
+12;18;3;
+13;21;3;
+14;23;3;
+15;23;1;
