+++ results/archiv/first_timetable.csv
@@ -1,0 +1,19 @@
+;1;;;2;;;3;;;4;;;5;;;6;;;7;;;8;;;9;;;10;;;11;;;12;;;13;;;14;;;15;;;16;;;17;;;18;;;19;;;20;;;21;;;22;;;23;;;24;;;25;;;26;;;27;;;28;;
+;1;2;3;1;2;3;1;2;3;1;2;3;1;2;3;1;2;3;1;2;3;1;2;3;1;2;3;1;2;3;1;2;3;1;2;3;1;2;3;1;2;3;1;2;3;1;2;3;1;2;3;1;2;3;1;2;3;1;2;3;1;2;3;1;2;3;1;2;3;1;2;3;1;2;3;1;2;3;1;2;3;1;2;3
+Nurse_1;1;0;0;0;1;0;1;0;0;1;0;0;1;0;0;1;0;0;1;0;0;0;0;1;0;1;0;0;1;0;1;0;0;1;0;0;0;0;1;0;0;1;0;1;0;0;0;1;0;1;0;0;1;0;1;0;0;1;0;0;1;1;0;1;0;0;1;0;0;0;1;0;0;0;1;0;1;0;1;0;0;0;0;1
+Nurse_2;1;0;0;0;1;0;1;0;0;1;0;0;0;1;0;1;0;0;1;1;1;0;0;1;0;0;1;0;1;0;1;0;0;1;0;0;0;1;0;1;0;0;0;1;0;0;0;1;0;0;1;0;1;0;1;0;0;1;0;0;1;0;0;0;0;1;0;0;1;0;1;0;1;0;0;0;1;0;1;0;0;1;0;0
+Nurse_3;0;1;0;1;0;0;0;0;1;0;1;0;0;1;0;0;1;0;0;0;1;0;1;0;0;0;1;0;0;1;0;0;1;0;1;0;1;0;0;0;1;0;1;0;0;1;0;0;0;0;1;0;0;1;0;0;1;0;0;1;0;0;1;0;0;1;0;1;0;1;0;0;1;0;0;0;0;1;0;1;0;0;1;0
+Nurse_4;0;1;0;0;1;0;1;0;0;0;0;1;0;1;0;0;1;0;0;1;0;1;0;0;0;0;1;0;0;1;0;1;0;0;1;0;0;1;0;1;0;0;0;0;1;0;0;1;0;0;1;0;0;1;0;1;0;1;0;0;1;0;0;0;0;1;0;0;1;0;1;0;1;0;0;1;0;0;0;1;0;1;0;0
+Nurse_5;1;0;0;0;0;1;0;1;0;0;1;0;1;0;1;1;0;0;1;0;0;0;1;0;0;1;0;0;1;0;0;0;1;0;0;1;0;0;1;0;1;0;1;0;0;0;1;0;0;1;0;0;1;0;1;0;0;0;0;1;0;0;1;0;1;0;0;1;0;0;0;1;0;1;0;0;0;1;0;0;1;0;1;0
+Nurse_6;0;1;0;1;0;0;0;0;1;0;1;0;0;1;0;0;1;0;0;0;1;0;1;0;0;0;1;0;0;1;0;0;1;0;1;0;1;0;0;0;1;0;1;0;0;1;0;0;0;0;1;0;0;1;0;0;1;0;0;1;0;0;1;0;0;1;0;1;0;1;0;0;1;0;0;0;0;1;0;1;0;0;1;0
+Nurse_7;1;0;0;0;1;0;1;0;0;1;0;0;0;1;0;1;0;0;1;0;0;0;0;1;0;0;1;0;1;0;1;0;0;1;0;0;0;1;0;1;0;0;0;1;0;0;0;1;0;0;1;0;1;0;1;0;0;1;0;0;1;0;0;0;0;1;0;0;1;0;1;0;1;0;1;0;1;0;1;0;0;1;0;0
+Nurse_8;0;0;1;0;0;1;0;1;0;1;0;0;1;0;0;1;0;0;1;0;0;0;0;1;0;1;0;0;1;0;1;0;0;0;0;1;0;0;1;0;0;1;0;1;0;0;1;0;0;1;0;1;0;0;1;0;0;0;1;0;0;1;0;1;0;0;1;0;0;0;0;1;0;0;1;0;1;0;1;0;0;0;0;1
+Nurse_9;0;1;0;0;1;0;1;0;0;0;0;1;0;1;0;0;1;0;0;1;0;1;0;0;0;0;1;0;0;1;0;1;0;1;1;0;0;1;0;1;0;0;0;0;1;0;0;1;0;0;1;0;0;1;0;1;0;1;0;0;1;0;0;0;0;1;0;0;1;0;1;0;1;0;0;1;0;0;0;1;0;1;0;0
+Nurse_10;0;1;0;0;1;0;1;0;0;0;0;1;0;1;0;0;1;0;0;1;1;1;0;0;0;0;1;0;0;1;0;1;0;0;1;0;0;1;0;1;0;0;0;0;1;0;0;1;0;0;1;0;0;1;0;1;0;0;1;0;1;0;0;0;0;1;0;0;1;0;1;0;1;0;0;1;0;0;0;1;0;1;0;0
+Nurse_11;0;0;0;0;0;0;0;0;1;0;0;0;0;0;1;0;0;1;0;0;0;0;0;1;0;0;0;1;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;1;0;1;0;1;0;0;0;0;0;0;0;0;0;0;1;0;1;0;0;0;0;0;0;0;1;0;0;0;1;0;0;0;0;1;0;0;0;0;0
+Nurse_12;1;0;0;0;0;0;1;0;0;1;0;0;1;0;0;1;0;0;1;0;0;0;0;1;0;1;0;0;1;0;1;0;0;1;0;0;0;0;1;0;0;1;0;1;0;1;0;0;0;1;0;0;1;0;1;0;0;1;0;0;0;1;0;1;0;0;1;0;0;0;0;0;0;0;1;0;1;0;1;0;0;0;0;1
+Nurse_13;0;1;0;1;0;0;0;0;1;0;1;0;0;0;1;0;0;1;0;1;0;1;1;0;1;0;0;1;0;0;0;0;1;0;1;0;1;0;0;0;1;0;1;0;0;1;0;0;1;0;0;0;0;1;0;0;1;0;0;1;0;0;1;0;1;0;0;1;0;1;0;0;0;1;0;1;0;0;0;0;1;0;1;0
+Nurse_14;0;0;1;0;0;0;0;0;0;0;0;0;0;0;1;0;0;1;0;1;0;1;0;0;1;0;0;1;0;0;0;0;0;0;0;0;0;1;0;0;0;0;0;0;0;0;0;0;1;0;0;1;0;0;0;1;0;0;1;0;0;1;0;0;1;0;0;0;0;0;0;0;0;1;0;1;0;0;0;0;1;0;0;0
+Nurse_15;0;0;1;0;0;1;0;1;0;1;0;0;1;0;0;1;0;0;1;0;1;0;0;1;0;1;0;0;1;0;1;0;0;0;0;1;0;0;1;0;0;1;0;1;0;0;1;0;0;1;0;1;0;0;1;0;0;0;1;0;0;1;0;1;0;0;1;0;0;0;0;1;0;0;1;0;1;0;1;0;0;0;0;1
+Nurses_on_shift;5;6;3;3;6;3;7;3;4;6;4;3;5;7;4;7;5;3;7;6;5;5;4;7;2;5;7;3;7;5;6;3;4;5;6;3;3;6;5;5;4;4;4;6;4;4;4;6;3;5;7;3;5;6;7;4;3;6;4;5;6;6;4;4;3;7;4;4;5;4;6;3;7;4;5;5;6;3;7;5;3;5;4;4
+Demanded_nurses;4;3;2;3;4;2;3;3;4;4;3;2;3;2;2;3;2;3;2;2;2;4;2;2;2;2;4;3;4;2;4;2;4;2;3;2;2;3;4;2;4;2;2;3;4;4;2;3;3;2;4;3;2;2;3;3;3;2;4;3;3;2;3;3;3;4;3;3;2;4;3;3;3;4;3;3;2;3;2;2;3;2;2;2
